Try using watchdog timer to fix frozen AQI display
After some long period of time, the display stops updating. And I don't notice for days.
My hypothesis is that somewhere in the stack (probably `fetch()` or `time.sleep()`)
there's a piece of code that's not handling a wrap-around of a 32 bit tick value. I saw
a comment that it'll be ~49.5 days, and that seems like it's plausibly the frequency that
I see this.

I briefly went looking for that bug, but I don't know the system well enough to find it.
So I considered adding "check ticks and then hard reset" logic to this loop, but stumbled
onto the watchdog, which I hope will be even more likely to fix my issues - even if my
hypothesis of the root cause is wrong.