Remove watchdog, use time.monotonic to reset every 24 days
The watchdog on the SAMD51 is limited to 16 seconds, which won't work for me at this
high level.
https://github.com/adafruit/circuitpython/blob/e39fbf1b26b4fd3b66313e51ccc3db0eba7bd58a/ports/atmel-samd/common-hal/watchdog/WatchDogTimer.c#L77

So I went looking for code that uses a 32 bit tick value instead, assuming that's the root
cause. I wasn't able to find a smoking gun. However, I _have_ added an "every 24 days,
reset the microcontroller" loop.

`supervisor.ticks_ms()` is set up for wrapping every (2^29)ms, and it *starts* so that
it'll wrap approximately 65s after power-on, to help find issues around overflow.

`time.monotonic` exposes the un-manipulated, un-wrapped ticks value. Meaning that _if_
this is caused by older code that's broken for wrap around, this should keep the underlying
ticks from ever getting large enough to exceed `uint32_t`

On the other hand, if this is something that's using `supervisor.ticks_ms`, it won't do
anything helpful.

There don't seem to be very many places that use the 32b ticks value, it looks like it's
mostly 64b.